Now keeping track of all stats in-node.
This changes the mcts nodes so that their winrate is held in their
parent rather than in the node itself, so that we can do UCT+RAVE
weighting on possible moves without having a node already in the tree.
This turns out to be pretty important for exploration after choosing
the initial random moves.

Also some slight tweaks to the debugger.